Extract networkview
